+++ src/module/transaction/screen/SendSetAmountScreen/SendSetAmountScreen.tsx
@@ -9,7 +9,7 @@ import { useTranslate } from "module/common/hook/useTranslate";
 import WalletAssetSelect from "module/wallet/component/input/WalletAssetSelect/WalletAssetSelect";
 import AssetAmountTextField from "module/transaction/component/input/AssetAmountTextField/AssetAmountTextField";
 import { useState } from "react";
-import { Asset } from "module/wallet/component/input/WalletAssetSelect/WalletAssetSelect.types";
+import { Asset } from "module/wallet/wallet.types";
 import { AssetType } from "module/wallet/wallet.types";
 
 export type SendAmountAndMessageResult = Pick<SendState, "amount" | "asset">;
@@ -23,6 +23,7 @@ const SendSetAmountScreen = (): JSX.Element => {
     const [sendState, setSendState] = useRecoilState(sendRecoilState);
     /**
      * asset will never be undefined because by default sendRoilState has it defined
+     * (check the recoil defaultState of sendState)
      */
     const [asset, setAsset] = useState<Asset | undefined>(sendState.asset);
     const translate = useTranslate();
