+++ src/module/wallet/component/input/WalletAssetSelect/WalletAssetSelect.types.ts
@@ -1,13 +1,6 @@
 import { CoreFormControlledComponentProps, FormControlledComponentProps, LabelProps } from "@peersyst/react-native-components";
-import { AssetType } from "module/wallet/wallet.types";
-import { NftToken, Token } from "near-peersyst-sdk";
+import { Asset } from "module/wallet/wallet.types";
 
 export type WalletAssetSelectProps = FormControlledComponentProps<CoreFormControlledComponentProps<Asset | undefined, LabelProps>> & {
     index: number;
 };
-
-export interface Asset {
-    type: AssetType;
-    nft?: NftToken;
-    ft?: Token;
-}
